moving Yup validation to schemas/login.js file
