+++ src/components/LoginForm/LoginFormContainer.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import axios from 'axios'
-import * as Yup from 'yup'
 import { useNavigate } from 'react-router-dom'
 import LoginForm from './LoginForm'
+import validationSchema from '../../schemas/login';
 
 const LoginFormContainer = () => {
   const [errorStatus, setErrorStatus] = useState(null)
@@ -30,12 +30,6 @@ const LoginFormContainer = () => {
     email: '',
     password: '',
   }
-  const validationSchema = Yup.object({
-    email: Yup.string().email('Dirección de mail inválida').required('Obligatorio'),
-    password: Yup.string()
-      .min(6, 'Debe tener al menos 6 caracteres')
-      .required('Obligatorio'),
-  })
 
   return (
     <LoginForm
